Feat: test cmake build

+++ src/net/EventLoopThreadPool.cpp
@@ -13,7 +13,7 @@ EventLoopThreadPool::~EventLoopThreadPool() {
 
 void EventLoopThreadPool::start(const ThreadInitCallback &cb) {
     _loopPtr->assertInLoopThread();
-    for (int i = 0; i < _threadNumbers; ++i) {
+    for (uint64_t i = 0; i < _threadNumbers; ++i) {
         EventLoopThread *th = new EventLoopThread(cb);
         _threads.push_back(std::unique_ptr<EventLoopThread>(th));
         _loops.push_back(th->startLoop());
